Rename recyclerv.interfaces package to interfaces
