+++ app/src/main/java/com/shapeapp/shape/ui/recyclerviews/adapters/SmallCardRecyclerViewAdapter.kt
@@ -9,7 +9,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.shapeapp.shape.R
 import com.shapeapp.shape.data.entities.Card
-import com.shapeapp.shape.ui.recyclerviews.recyclerviewinterfaces.RecyclerViewCardClickListener
+import com.shapeapp.shape.ui.recyclerviews.interfaces.RecyclerViewCardClickListener
 
 /**
  * Feeds [RecyclerView] with [Card] data
